+++ crates/signum/src/chsets/cache.rs
@@ -1,6 +1,10 @@
+//! # Implementation of a charset cache
+
 use std::{collections::HashMap, fs::DirEntry, path::Path, path::PathBuf};
 
-use signum::chsets::{
+use log::{info, warn};
+
+use crate::chsets::{
     editor::ESet,
     editor::OwnedESet,
     encoding::{p_mapping_file, Mapping},
@@ -21,7 +25,7 @@ fn find_font_file(cset_folder: &Path, name: &str, extension: &str) -> Option<Pat
     let mut dir_iter = match std::fs::read_dir(cset_folder) {
         Ok(i) => i,
         Err(e) => {
-            println!("Could not find CHSET folder: {}", e);
+            warn!("Could not find CHSET folder: {}", e);
             return None;
         }
     };
@@ -47,12 +51,12 @@ fn load_printer_font(editor_cset_file: &Path, pk: PrinterKind) -> Option<OwnedPS
     let printer_cset_file = editor_cset_file.with_extension(extension);
     match OwnedPSet::load(&printer_cset_file, pk) {
         Ok(pset) => {
-            println!("Loaded printer font file '{}'", printer_cset_file.display());
+            info!("Loaded printer font file '{}'", printer_cset_file.display());
             Some(pset)
         }
         Err(LoadError::Io(_)) => None,
         Err(LoadError::Parse(f)) => {
-            println!("Failed to parse {} printer font: {}", extension, f);
+            warn!("Failed to parse {} printer font: {}", extension, f);
             None
         }
     }
@@ -64,12 +68,12 @@ fn load_mapping_file(editor_cset_file: &Path) -> Option<Mapping> {
         let input = std::fs::read_to_string(&cset_mapping_file).unwrap();
         match p_mapping_file(&input) {
             Ok(mapping) => {
-                eprintln!("Loaded cset mapping file '{}'", cset_mapping_file.display());
+                info!("Loaded cset mapping file '{}'", cset_mapping_file.display());
                 Some(mapping)
             }
             Err(err) => {
-                eprintln!(
-                    "[cli::font::cache] Failed to parse mapping file '{}': {}",
+                warn!(
+                    "Failed to parse mapping file '{}': {}",
                     cset_mapping_file.display(),
                     err
                 );
@@ -77,8 +81,8 @@ fn load_mapping_file(editor_cset_file: &Path) -> Option<Mapping> {
             }
         }
     } else {
-        eprintln!(
-            "[cli::font::cache] missing mapping for font '{}",
+        warn!(
+            "Missing mapping for font '{}",
             editor_cset_file.file_stem().unwrap().to_string_lossy()
         );
         None
@@ -88,29 +92,34 @@ fn load_mapping_file(editor_cset_file: &Path) -> Option<Mapping> {
 fn load_editor_font(editor_cset_file: &Path) -> Option<OwnedESet> {
     match OwnedESet::load(&editor_cset_file) {
         Ok(eset) => {
-            println!("Loaded editor font file '{}'", editor_cset_file.display());
+            info!("Loaded editor font file '{}'", editor_cset_file.display());
             Some(eset)
         }
         Err(LoadError::Parse(e)) => {
-            println!(
-                "Failed to parse editor font file {}",
-                editor_cset_file.display()
+            info!(
+                "Failed to parse editor font file {}
+                Are you sure this is a valid Signum! editor font?
+                Error: {}",
+                editor_cset_file.display(),
+                e
             );
-            println!("Are you sure this is a valid Signum! editor font?");
-            println!("Error: {}", e);
             None
         }
         Err(LoadError::Io(e)) => {
-            println!(
-                "Failed to load editor font file {}",
-                editor_cset_file.display()
+            warn!(
+                "Failed to load editor font file {}
+                Error: {}",
+                editor_cset_file.display(),
+                e
             );
-            println!("Error: {}", e);
             None
         }
     }
 }
 
+/// Holds variants of a charset
+///
+/// This structure holds different representations (e.g. Bitmaps for different printer kinds) of the same character set.
 pub struct CSet {
     name: String,
     l30: Option<OwnedPSet>,
@@ -122,12 +131,19 @@ pub struct CSet {
 
 #[rustfmt::skip]
 impl CSet {
+    /// The the name of the character set
     pub fn name(&self) -> &str { &self.name }
+    /// Get the unicode mapping
     pub fn map(&self) -> Option<&Mapping> { self.map.as_ref() }
+    /// Get the laser printer bitmaps
     pub fn l30(&self) -> Option<&PSet<'static>> { self.l30.as_deref() }
+    /// Get the 24-needle printer bitmaps
     pub fn p24(&self) -> Option<&PSet<'static>> { self.p24.as_deref() }
+    /// Get the 9-needle printer bitmaps
     pub fn p09(&self) -> Option<&PSet<'static>> { self.p09.as_deref() }
+    /// Get the editor bitmaps
     pub fn e24(&self) -> Option<&ESet<'static>> { self.e24.as_deref() }
+    /// Get the bitmaps for the specified printer kind
     pub fn printer(&self, pk: PrinterKind) -> Option<&PSet<'static>> {
         match pk {
             PrinterKind::Needle9 => self.p09.as_deref(),
@@ -137,37 +153,44 @@ impl CSet {
     }
 }
 
-pub struct FontCache {
+/// A simple cache for charsets
+pub struct ChsetCache {
     chsets_folder: PathBuf,
     chsets: Vec<CSet>,
     names: HashMap<String, usize>,
 }
 
-impl FontCache {
+impl ChsetCache {
+    /// Create a new instance
     pub fn new(chsets_folder: PathBuf) -> Self {
-        FontCache {
+        ChsetCache {
             chsets: Vec::with_capacity(8),
             names: HashMap::new(),
             chsets_folder,
         }
     }
 
+    /// Get the slice of all charsets
     pub fn chsets(&self) -> &[CSet] {
         &self.chsets
     }
 
+    /// Get a specific printer charset
     pub fn pset(&self, pk: PrinterKind, index: usize) -> Option<&PSet<'static>> {
         self.cset(index).and_then(|cset| cset.printer(pk))
     }
 
+    /// Get a specific editor charset
     pub fn eset(&self, index: usize) -> Option<&ESet<'static>> {
         self.cset(index).and_then(CSet::e24)
     }
 
+    /// Get a specific charset
     pub fn cset(&self, index: usize) -> Option<&CSet> {
         self.chsets.get(index)
     }
 
+    /// Load a character set
     pub fn load_cset(&mut self, name: &str) -> Option<usize> {
         if let Some(index) = self.names.get(name) {
             return Some(*index);
@@ -176,7 +199,7 @@ impl FontCache {
         let editor_cset_file = match find_font_file(&self.chsets_folder, name, "E24") {
             Some(f) => f,
             None => {
-                println!("Editor font for `{}` not found!", name);
+                warn!("Editor font for `{}` not found!", name);
                 return None;
             }
         };