+++ src/cli/font/mod.rs
@@ -17,7 +17,6 @@ use std::{
     path::{Path, PathBuf},
 };
 
-pub mod cache;
 pub mod ps;
 
 use ps::write_ls30_ps_bitmap;