+++ src/cli/sdoc/ps.rs
@@ -2,15 +2,15 @@ use std::{fs::File, io::BufWriter, io::Write, path::Path};
 
 use color_eyre::eyre::{self, eyre};
 use sdo_ps::out::PSWriter;
-use signum::chsets::FontKind;
+use signum::chsets::{cache::ChsetCache, FontKind};
 
-use crate::cli::font::{cache::FontCache, ps::write_ls30_ps_bitmap};
+use crate::cli::font::ps::write_ls30_ps_bitmap;
 
 use super::{ps_proc::prog_dict, Document};
 
 fn output_ps_writer(
     doc: &Document,
-    fc: &FontCache,
+    fc: &ChsetCache,
     pw: &mut PSWriter<impl Write>,
 ) -> eyre::Result<()> {
     let pd = doc
@@ -171,7 +171,7 @@ fn output_ps_writer(
     Ok(())
 }
 
-pub fn output_postscript(doc: &Document, fc: &FontCache) -> eyre::Result<()> {
+pub fn output_postscript(doc: &Document, fc: &ChsetCache) -> eyre::Result<()> {
     if doc.opt.out.as_deref() == Some(Path::new("-")) {
         println!("----------------------------- PostScript -----------------------------");
         let mut pw = PSWriter::new();