+++ crates/pdf/CHANGES.md
@@ -1,5 +1,13 @@
-# 2020-11-24
+# Changelog
+
+## 2021-02-26
+
+- Switched to `CharProc` -> `ASCII85Stream`
+- Added optional image metadata to `ASCII85Stream`
+- Added `ColorSpace` enum
+
+## 2020-11-24
 
 - Fix EOL in classic xref section to be `SP LF`
 
-# 2020-11-23 (v0.1)
+## 2020-11-23 (v0.1)